load animations only once
