+++ C7/Civ3AnimData.cs
@@ -1,17 +1,17 @@
 
-// Civ3AnimData's purpose is to store the data associated with Civ 3 animations, for example the contents of each folder in Art/Units. It does lazy
+// AnimationManager's purpose is to store the data associated with Civ 3 animations, for example the contents of each folder in Art/Units. It does lazy
 // loading & memoization so each file is loaded only when it's needed and then stored so it is only ever loaded once per game. The main (and only)
-// instance of Civ3AnimData is kept in Game, AnimationTracker holds a reference to it.
+// instance of AnimationManager is kept in Game, AnimationTracker holds a reference to it.
 
 // It would be nice to load this data close to where it's used, e.g., have UnitLayer load the FlicSheets, instead of putting all the loading code in
-// one detached class like this. That's how things originally worked but I created Civ3AnimData to solve two issues:
+// one detached class like this. That's how things originally worked but I created AnimationManager to solve two issues:
 // 1. AnimationTracker and UnitLayer both need to load the unit INIs. So we either have a common place to store the INIs or duplication of work, and I
 // think the former is the better choice.
 // 2. AnimationTracker needs to know the duration of animations, which awkwardly cannot be determined based on the INI files alone. In order to know
 // the duration of an anim you must know how many frames it has, and the only way to know that is to read its flic file.
 
 // The intended usage is to access the animation data through a Civ3Anim object obtained through the "forUnit" or "forEffect" methods. For example:
-//   civ3AnimData.forUnit("Warrior", MapUnit.AnimatedAction.FORTIFY).playSound()
+//   AnimationManager.forUnit("Warrior", MapUnit.AnimatedAction.FORTIFY).playSound()
 // To play the warrior's foritfy sound effect.
 
 using System;
@@ -20,9 +20,9 @@ using Godot;
 using IniParser;
 using IniParser.Model;
 using C7GameData;
+using ConvertCiv3Media;
 
-public partial class Civ3AnimData
-{
+public partial class AnimationManager {
 
 	public static string BaseAnimationKey(string unitName, MapUnit.AnimatedAction action) {
 		return String.Format("{0}_{1}", unitName, action.ToString());
@@ -42,17 +42,20 @@ public partial class Civ3AnimData
 
 	private AudioStreamPlayer audioPlayer;
 
-	public Civ3AnimData(AudioStreamPlayer audioPlayer)
-	{
-		this.audioPlayer = audioPlayer;
-	}
+	public SpriteFrames spriteFrames;
+	public SpriteFrames tintFrames;
 
 	private Dictionary<string, IniData> iniDatas = new Dictionary<string, IniData>();
 
-	public IniData getINIData(string pathKey)
-	{
+	public AnimationManager(AudioStreamPlayer audioPlayer) {
+		this.audioPlayer = audioPlayer;
+		this.spriteFrames = new SpriteFrames();
+		this.tintFrames = new SpriteFrames();
+	}
+
+	public IniData getINIData(string pathKey) {
 		IniData tr;
-		if (! iniDatas.TryGetValue(pathKey, out tr)) {
+		if (!iniDatas.TryGetValue(pathKey, out tr)) {
 			string fullPath = Util.Civ3MediaPath(pathKey);
 			tr = (new FileIniDataParser()).ReadFile(fullPath);
 			iniDatas.Add(pathKey, tr);
@@ -60,31 +63,82 @@ public partial class Civ3AnimData
 		return tr;
 	}
 
-	public IniData getUnitINIData(string unitTypeName)
-	{
+	// Looks up the name of the flic file associated with a given action in an animation INI. If there is no flic file listed for the action,
+	// returns instead the file name for the default action, and if that's missing too, throws an exception.
+	public string getFlicFileName(IniData iniData, MapUnit.AnimatedAction action) {
+		string fileName = iniData["Animations"][action.ToString()];
+		if (fileName != null && fileName != "") {
+			return fileName;
+		} else if (action != MapUnit.AnimatedAction.DEFAULT) {
+			return getFlicFileName(iniData, MapUnit.AnimatedAction.DEFAULT);
+		} else {
+			throw new Exception("Missing default animation"); // TODO: Add the INI's file name to the error message
+		}
+	}
+
+	public IniData getUnitINIData(string unitTypeName) {
 		return getINIData(String.Format("Art/Units/{0}/{0}.INI", unitTypeName));
 	}
 
-	// Looks up the name of the flic file associated with a given action in an animation INI. If there is no flic file listed for the action,
-	// returns instead the file name for the default action, and if that's missing too, throws an exception.
-	public string getFlicFileName(IniData iniData, MapUnit.AnimatedAction action)
-	{
-		string fileName = iniData["Animations"][action.ToString()];
-		if ((fileName != null) && (fileName != ""))
-			return fileName;
-		else if (action != MapUnit.AnimatedAction.DEFAULT)
-			return getFlicFileName(iniData, MapUnit.AnimatedAction.DEFAULT);
-		else
-			throw new Exception("Missing default animation"); // TODO: Add the INI's file name to the error message
+	public string getUnitFlicFilepath(UnitPrototype unit, MapUnit.AnimatedAction action) {
+		string directory = String.Format("Art/Units/{0}", unit.name);
+		IniData ini = getUnitINIData(unit.name);
+		string filename = getFlicFileName(ini, action);
+		return directory.PathJoin(filename);
+	}
+
+	// The flic loading code parses the animations into a 2D array, where each row is an animation
+	// corresponding to a tile direction. flicRowToAnimationDirection maps row number -> direction.
+	private static TileDirection flicRowToAnimationDirection(int row) {
+		switch (row) {
+			case 0: return TileDirection.SOUTHWEST;
+			case 1: return TileDirection.SOUTH;
+			case 2: return TileDirection.SOUTHEAST;
+			case 3: return TileDirection.EAST;
+			case 4: return TileDirection.NORTHEAST;
+			case 5: return TileDirection.NORTH;
+			case 6: return TileDirection.NORTHWEST;
+			case 7: return TileDirection.WEST;
+		}
+		return TileDirection.NORTH;
+	}
+
+	public static (SpriteFrames, SpriteFrames) loadFlicAnimation(string path, string name, ref SpriteFrames frames, ref SpriteFrames tint) {
+		Flic flic = Util.LoadFlic(path);
+
+		for (int row = 0; row < flic.Images.GetLength(0); row++) {
+			string direction = flicRowToAnimationDirection(row).ToString();
+			string animationName = name + "_" + direction;
+			frames.AddAnimation(animationName);
+			tint.AddAnimation(animationName);
+
+			for (int col = 0; col < flic.Images.GetLength(1); col++) {
+				byte[] frame = flic.Images[row,col];
+				(ImageTexture bl, ImageTexture tl) = Util.LoadTextureFromFlicData(frame, flic.Palette, flic.Width, flic.Height);
+				frames.AddFrame(animationName, bl, 0.5f); // TODO: frame duration is controlled by .ini
+				tint.AddFrame(animationName, tl, 0.5f);   // TODO: frame duration is controlled by .ini
+			}
+		}
+		return (frames, tint);
+	}
+
+	public bool LoadAnimation(UnitPrototype unit, MapUnit.AnimatedAction action) {
+		string name = BaseAnimationKey(unit.name, action);
+		string testName = AnimationKey(name, TileDirection.NORTH);
+		if (spriteFrames.HasAnimation(testName) && tintFrames.HasAnimation(testName)) {
+			return false;
+		}
+		string filepath = getUnitFlicFilepath(unit, action);
+		loadFlicAnimation(filepath, name, ref this.spriteFrames, ref this.tintFrames);
+		return true;
 	}
 
 	private Dictionary<string, Util.FlicSheet> flicSheets = new Dictionary<string, Util.FlicSheet>();
 
-	public Util.FlicSheet getFlicSheet(string rootPath, IniData iniData, MapUnit.AnimatedAction action)
-	{
+	public Util.FlicSheet getFlicSheet(string rootPath, IniData iniData, MapUnit.AnimatedAction action) {
 		Util.FlicSheet tr;
 		string pathKey = rootPath + "/" + getFlicFileName(iniData, action);
-		if (! flicSheets.TryGetValue(pathKey, out tr)) {
+		if (!flicSheets.TryGetValue(pathKey, out tr)) {
 			(tr, _) = Util.loadFlicSheet(pathKey);
 			flicSheets.Add(pathKey, tr);
 		}
@@ -93,13 +147,12 @@ public partial class Civ3AnimData
 
 	private Dictionary<string, AudioStreamWav> wavs = new Dictionary<string, AudioStreamWav>();
 
-	public void playSound(string rootPath, IniData iniData, MapUnit.AnimatedAction action)
-	{
+	public void playSound(string rootPath, IniData iniData, MapUnit.AnimatedAction action) {
 		string fileName = iniData["Sound Effects"][action.ToString()];
 		if (fileName.EndsWith(".WAV", StringComparison.CurrentCultureIgnoreCase)) {
 			AudioStreamWav wav;
 			var pathKey = rootPath + "/" + fileName;
-			if (! wavs.TryGetValue(pathKey, out wav)) {
+			if (!wavs.TryGetValue(pathKey, out wav)) {
 				wav = Util.LoadWAVFromDisk(Util.Civ3MediaPath(pathKey));
 				wavs.Add(pathKey, wav);
 			}
@@ -108,32 +161,28 @@ public partial class Civ3AnimData
 		}
 	}
 
-	public Civ3Anim forUnit(string unitTypeName, MapUnit.AnimatedAction action)
-	{
-		return new Civ3Anim(this, unitTypeName, action);
+	public C7Animation forUnit(UnitPrototype unit, MapUnit.AnimatedAction action) {
+		return new C7Animation(this, unit, action);
 	}
 
-	public Civ3Anim forEffect(AnimatedEffect effect)
-	{
-		return new Civ3Anim(this, effect);
+	public C7Animation forEffect(AnimatedEffect effect) {
+		return new C7Animation(this, effect);
 	}
 }
 
-public partial class Civ3Anim
-{
-	public Civ3AnimData civ3AnimData  { get; private set; }
+public partial class C7Animation {
+	public AnimationManager animationManager { get; private set; }
 	public string folderPath { get; private set; } // For example "Art/Units/Warrior" or "Art/Animations/Trajectory"
 	public string iniFileName { get; private set; }
-	private string unitType;
+	private UnitPrototype unit;
 	public MapUnit.AnimatedAction action { get; private set; }
 
-	public Civ3Anim(Civ3AnimData civ3AnimData, string unitTypeName, MapUnit.AnimatedAction action)
-	{
-		this.civ3AnimData = civ3AnimData;
-		this.folderPath = "Art/Units/" + unitTypeName;
-		this.iniFileName = unitTypeName + ".ini";
+	public C7Animation(AnimationManager civ3AnimData, UnitPrototype unit, MapUnit.AnimatedAction action) {
+		this.animationManager = civ3AnimData;
+		this.folderPath = "Art/Units/" + unit.name;
+		this.iniFileName = unit.name + ".ini";
 		this.action = action;
-		this.unitType = unitTypeName;
+		this.unit = unit;
 	}
 
 	public static readonly Dictionary<AnimatedEffect, string> effectCategories = new Dictionary<AnimatedEffect, string>
@@ -156,42 +205,30 @@ public partial class Civ3Anim
 		{ AnimatedEffect.WaterMiss, "water miss.ini" }
 	};
 
-	public Civ3Anim(Civ3AnimData civ3AnimData, AnimatedEffect effect)
-	{
-		this.civ3AnimData = civ3AnimData;
+	public C7Animation(AnimationManager civ3AnimData, AnimatedEffect effect) {
+		this.animationManager = civ3AnimData;
 		this.folderPath = "Art/Animations/" + effectCategories[effect];
 		this.iniFileName = effectINIFileNames[effect];
 		this.action = MapUnit.AnimatedAction.DEATH;
 	}
 
-	public IniData getINIData()
-	{
-		return civ3AnimData.getINIData(folderPath + "/" + iniFileName);
+	public IniData getINIData() {
+		return animationManager.getINIData(folderPath + "/" + iniFileName);
 	}
 
-	public Util.FlicSheet getFlicSheet()
-	{
-		return civ3AnimData.getFlicSheet(folderPath, getINIData(), action);
+	public Util.FlicSheet getFlicSheet() {
+		return animationManager.getFlicSheet(folderPath, getINIData(), action);
 	}
 
-	public void loadSpriteAnimation(AnimatedSprite2D animation) {
-		string baseName = Civ3AnimData.BaseAnimationKey(this.unitType, this.action);
-		if (animation.SpriteFrames.HasAnimation(Civ3AnimData.AnimationKey(baseName, TileDirection.NORTH))) {
-			// pick a random direction to check, since all directions are loaded from a single flic file
-			return;
-		}
-		string flicPath = this.folderPath + "/" + this.civ3AnimData.getFlicFileName(getINIData(), this.action);
-		SpriteFrames spriteFrames = animation.SpriteFrames; // ref semantics feel kinda dumb since it's not even necessary in C#
-		Util.loadFlicAnimation(flicPath, baseName, ref spriteFrames);
+	public void loadSpriteAnimation() {
+		this.animationManager.LoadAnimation(this.unit, this.action);
 	}
 
-	public void playSound()
-	{
-		civ3AnimData.playSound(folderPath, getINIData(), action);
+	public void playSound() {
+		animationManager.playSound(folderPath, getINIData(), action);
 	}
 
-	public double getDuration()
-	{
+	public double getDuration() {
 		Util.FlicSheet flicSheet = getFlicSheet();
 		double frameCount = flicSheet.indices.GetWidth() / flicSheet.spriteWidth;
 		return frameCount / 20.0; // Civ 3 anims often run at 20 FPS   TODO: Do they all? How could we tell? Is it exactly 20 FPS?
