+++ C7/MapView.cs
@@ -466,414 +466,6 @@ public partial class GridLayer : LooseLayer {
 	}
 }
 
-public partial class UnitLayer : LooseLayer {
-	private ImageTexture unitIcons;
-	private int unitIconsWidth;
-	private ImageTexture unitMovementIndicators;
-
-	// The unit animations, effect animations, and cursor are all drawn as children attached to the looseView but aren't created and attached in
-	// any particular order so we must use the ZIndex property to ensure they're properly layered.
-	public const int effectAnimZIndex = 150;
-	public const int unitAnimZIndex = 100;
-	public const int cursorZIndex = 50;
-
-	public UnitLayer()
-	{
-		var iconPCX = new Pcx(Util.Civ3MediaPath("Art/Units/units_32.pcx"));
-		unitIcons = PCXToGodot.getImageTextureFromPCX(iconPCX);
-		unitIconsWidth = (unitIcons.GetWidth() - 1) / 33;
-
-		var moveIndPCX = new Pcx(Util.Civ3MediaPath("Art/interface/MovementLED.pcx"));
-		unitMovementIndicators = PCXToGodot.getImageTextureFromPCX(moveIndPCX);
-	}
-
-	// Creates a plane mesh facing the positive Z-axis with the given shader attached. The quad is 1.0 units long on both sides,
-	// intended to be scaled to the appropriate size when used.
-	public static (ShaderMaterial, MeshInstance2D) createShadedQuad(Shader shader)
-	{
-		PlaneMesh mesh = new PlaneMesh();
-		mesh.SubdivideDepth = 1;
-		mesh.Orientation = PlaneMesh.OrientationEnum.Z;
-		mesh.Size = new Vector2(1, 1);
-
-		ShaderMaterial shaderMat = new ShaderMaterial();
-		shaderMat.Shader = shader;
-
-		MeshInstance2D meshInst = new MeshInstance2D();
-		meshInst.Material = shaderMat;
-		meshInst.Mesh = mesh;
-
-		return (shaderMat, meshInst);
-	}
-
-	public Color getHPColor(float fractionRemaining)
-	{
-		if (fractionRemaining >= (float)0.67) {
-			return Color.Color8(0, 255, 0);
-		} else if (fractionRemaining >= (float)0.34) {
-			return Color.Color8(255, 255, 0);
-		} else {
-			return Color.Color8(255, 0, 0);
-		}
-	}
-
-	// AnimationInstance represents an animation appearing on the screen. It's specific to a unit, action, and direction. AnimationInstances have
-	// two components: a ShaderMaterial and a MeshInstance2D. The ShaderMaterial runs the unit shader (created by UnitLayer.getShader) with all
-	// the parameters set to a particular texture, civ color, direction, etc. The MeshInstance2D is what's actually drawn by Godot, i.e., what's
-	// added to the node tree. AnimationInstances are only active for one frame at a time but they live as long as the UnitLayer. They are
-	// retrieved or created as needed by getBlankAnimationInstance during the drawing of units and are hidden & requeued for use at the beginning
-	// of each frame.
-
-	// should hold animation players instead of animations
-	public partial class AnimationInstance {
-
-		public AnimatedSprite2D sprite;
-		public AnimatedSprite2D spriteTint;
-		public ShaderMaterial material;
-
-		public void SetPosition(Vector2 position) {
-			this.sprite.Position = position;
-			this.spriteTint.Position = position;
-		}
-
-		public int GetNextFrameByProgress(string animation, float progress) {
-			// AnimatedSprite2D has a settable FrameProgress field, which I expected to
-			// update the current frame of the animation upon setting, but it did not
-			// when I tried it, so instead, calculate what the next frame should be
-			// based on the progress.
-			int frameCount = this.sprite.SpriteFrames.GetFrameCount(animation);
-			int nextFrame = (int)((float)frameCount * progress);
-			return nextFrame >= frameCount ? frameCount - 1 : (nextFrame < 0 ? 0 : nextFrame);
-		}
-
-		public void SetFrame(int frame) {
-			this.sprite.Frame = frame;
-			this.spriteTint.Frame = frame;
-		}
-
-		public AnimationInstance(LooseView looseView)
-		{
-			this.sprite = new AnimatedSprite2D();
-			SpriteFrames frames = new SpriteFrames();
-			this.sprite.SpriteFrames = frames;
-			this.sprite.ZIndex = unitAnimZIndex;
-
-			this.spriteTint = new AnimatedSprite2D();
-			SpriteFrames framesTint = new SpriteFrames();
-			this.spriteTint.SpriteFrames = framesTint;
-
-			this.material = new ShaderMaterial();
-			this.material.Shader = GD.Load<Shader>("res://tests/Tint.gdshader");
-			this.spriteTint.Material = this.material;
-			this.sprite.ZIndex = unitAnimZIndex;
-
-			looseView.AddChild(sprite);
-			looseView.AddChild(spriteTint);
-		}
-	}
-
-	private List<AnimationInstance> animInsts = new List<AnimationInstance>();
-	private int nextBlankAnimInst = 0;
-
-	// Returns the next unused AnimationInstance or creates & returns a new one if none are available.
-	public AnimationInstance getBlankAnimationInstance(LooseView looseView)
-	{
-		if (nextBlankAnimInst >= animInsts.Count) {
-			animInsts.Add(new AnimationInstance(looseView));
-		}
-		AnimationInstance inst = animInsts[nextBlankAnimInst];
-		nextBlankAnimInst++;
-		return inst;
-	}
-
-	// Sets the palette, indices, relSpriteSize, and spriteXY parameters on a ShaderMaterial to pick a sprite from a FlicSheet. relativeColumn
-	// varies between 0.0 for the first column and 1.0 for the last one.
-	public static void setFlicShaderParams(ShaderMaterial mat, Util.FlicSheet flicSheet, int row, float relativeColumn)
-	{
-		mat.SetShaderParameter("palette", flicSheet.palette);
-		mat.SetShaderParameter("indices", flicSheet.indices);
-
-		var indicesDims = new Vector2(flicSheet.indices.GetWidth(), flicSheet.indices.GetHeight());
-		var spriteSize = new Vector2(flicSheet.spriteWidth, flicSheet.spriteHeight);
-		mat.SetShaderParameter("relSpriteSize", spriteSize / indicesDims);
-
-		int spritesPerRow = flicSheet.indices.GetWidth() / flicSheet.spriteWidth;
-		int spriteColumn = (int)(relativeColumn * spritesPerRow);
-		if (spriteColumn >= spritesPerRow)
-			spriteColumn = spritesPerRow - 1;
-		else if (spriteColumn < 0)
-			spriteColumn = 0;
-		mat.SetShaderParameter("spriteXY", new Vector2(spriteColumn, row));
-	}
-
-	public void drawUnitAnimFrame(LooseView looseView, MapUnit unit, MapUnit.Appearance appearance, Vector2 tileCenter)
-	{
-		/*
-			This whole function is really hacky since I ported everything in the fastest way possible to get
-			animation working with Godot 4.
-		*/
-		var animOffset = MapView.cellSize * new Vector2(appearance.offsetX, appearance.offsetY);
-
-		// Need to move the sprites upward a bit so that their feet are at the center of the tile. I don't know if spriteHeight/4 is the right
-		var position = tileCenter + animOffset;// - new Vector2(0, flicSheet.spriteHeight / 4);
-
-		AnimationInstance inst = getBlankAnimationInstance(looseView);
-		looseView.mapView.game.civ3AnimData.forUnit(unit.unitType.name, appearance.action).loadSpriteAnimation(inst.sprite);
-		looseView.mapView.game.civ3AnimData.forUnit(unit.unitType.name, appearance.action).loadSpriteAnimation(inst.spriteTint);
-
-		inst.SetPosition(position);
-
-		string animName = Civ3AnimData.AnimationKey(unit.unitType, appearance.action, appearance.direction);
-		string animTintName = "TINT_" + animName;
-
-		var civColor = new Color(unit.owner.color);
-		int nextFrame = inst.GetNextFrameByProgress(animName, appearance.progress);
-		inst.SetFrame(nextFrame);
-
-		inst.material.SetShaderParameter("tintColor", new Vector3(civColor.R, civColor.G, civColor.B));
-		inst.sprite.Show();
-		inst.sprite.Animation = animName;
-
-		inst.sprite.Frame = nextFrame;
-		inst.spriteTint.Show();
-		inst.spriteTint.Animation = animTintName;
-	}
-
-	public void drawEffectAnimFrame(LooseView looseView, Civ3Anim anim, float progress, Vector2 tileCenter)
-	{
-		// var flicSheet = anim.getFlicSheet();
-		// var inst = getBlankAnimationInstance(looseView);
-		// setFlicShaderParams(inst.shaderMat, flicSheet, 0, progress);
-		// inst.shaderMat.SetShaderParameter("civColor", new Vector3(1, 1, 1));
-		// inst.meshInst.Position = tileCenter;
-		// inst.meshInst.Scale = new Vector2(flicSheet.spriteWidth, -1 * flicSheet.spriteHeight);
-		// inst.meshInst.ZIndex = effectAnimZIndex;
-	}
-
-	private Util.FlicSheet cursorFlicSheet;
-	private ShaderMaterial cursorMat = null;
-	private MeshInstance2D cursorMesh = null;
-
-	public void drawCursor(LooseView looseView, Vector2 position)
-	{
-		// Initialize cursor if necessary
-		if (cursorMesh == null) {
-			(cursorFlicSheet, _) = Util.loadFlicSheet("Art/Animations/Cursor/Cursor.flc");
-			(cursorMat, cursorMesh) = createShadedQuad(getCursorShader());
-			cursorMesh.Scale = new Vector2(cursorFlicSheet.spriteWidth, -1 * cursorFlicSheet.spriteHeight);
-			cursorMesh.ZIndex = cursorZIndex;
-			looseView.AddChild(cursorMesh);
-		}
-
-		const double period = 2.5; // TODO: Just eyeballing this for now. Read the actual period from the INI or something.
-		var repCount = (double)Time.GetTicksMsec() / 1000.0 / period;
-		var progress = (float)(repCount - Math.Floor(repCount));
-
-		setFlicShaderParams(cursorMat, cursorFlicSheet, 0, progress);
-		cursorMesh.Position = position;
-		cursorMesh.Show();
-	}
-
-	public override void onBeginDraw(LooseView looseView, GameData gameData)
-	{
-		// Reset animation instances
-		for (int n = 0; n < nextBlankAnimInst; n++) {
-			animInsts[n].sprite.Hide();
-			animInsts[n].spriteTint.Hide();
-		}
-		nextBlankAnimInst = 0;
-
-		// Hide cursor if it's been initialized
-		if (cursorMesh != null) {
-			cursorMesh.Hide();
-		}
-
-		looseView.mapView.game.updateAnimations(gameData);
-	}
-
-	// Returns which unit should be drawn from among a list of units. The list is assumed to be non-empty.
-	public MapUnit selectUnitToDisplay(LooseView looseView, List<MapUnit> units)
-	{
-		// From the list, pick out which units are (1) the strongest defender vs the currently selected unit, (2) the currently selected unit
-		// itself if it's in the list, and (3) any unit that is playing an animation that the player would want to see.
-		MapUnit bestDefender = units[0],
-			selected = null,
-			doingInterestingAnimation = null;
-		var currentlySelectedUnit = looseView.mapView.game.CurrentlySelectedUnit;
-		foreach (var u in units) {
-			if (u == currentlySelectedUnit)
-				selected = u;
-			if (u.HasPriorityAsDefender(bestDefender, currentlySelectedUnit))
-				bestDefender = u;
-			if (looseView.mapView.game.animTracker.getUnitAppearance(u).DeservesPlayerAttention())
-				doingInterestingAnimation = u;
-		}
-
-		// Prefer showing the selected unit, secondly show one doing a relevant animation, otherwise show the top defender
-		return selected != null ? selected : (doingInterestingAnimation != null ? doingInterestingAnimation : bestDefender);
-	}
-
-	public override void drawObject(LooseView looseView, GameData gameData, Tile tile, Vector2 tileCenter)
-	{
-		// First draw animated effects. These will always appear over top of units regardless of draw order due to z-index.
-		Civ3Anim tileEffect = looseView.mapView.game.animTracker.getTileEffect(tile);
-		if (tileEffect != null) {
-			(_, float progress) = looseView.mapView.game.animTracker.getCurrentActionAndProgress(tile);
-			drawEffectAnimFrame(looseView, tileEffect, progress, tileCenter);
-		}
-
-		if (tile.unitsOnTile.Count == 0) {
-			return;
-		}
-
-		var white = Color.Color8(255, 255, 255);
-
-		MapUnit unit = selectUnitToDisplay(looseView, tile.unitsOnTile);
-		MapUnit.Appearance appearance = looseView.mapView.game.animTracker.getUnitAppearance(unit);
-		Vector2 animOffset = new Vector2(appearance.offsetX, appearance.offsetY) * MapView.cellSize;
-
-		// If the unit we're about to draw is currently selected, draw the cursor first underneath it
-		if ((unit != MapUnit.NONE) && (unit == looseView.mapView.game.CurrentlySelectedUnit)) {
-			drawCursor(looseView, tileCenter + animOffset);
-		}
-
-		drawUnitAnimFrame(looseView, unit, appearance, tileCenter);
-
-		Vector2 indicatorLoc = tileCenter - new Vector2(26, 40) + animOffset;
-
-		int moveIndIndex = (!unit.movementPoints.canMove) ? 4 : ((unit.movementPoints.remaining >= unit.unitType.movement) ? 0 : 2);
-		Vector2 moveIndUpperLeft = new Vector2(1 + 7 * moveIndIndex, 1);
-		Rect2 moveIndRect = new Rect2(moveIndUpperLeft, new Vector2(6, 6));
-		var screenRect = new Rect2(indicatorLoc, new Vector2(6, 6));
-		looseView.DrawTextureRectRegion(unitMovementIndicators, screenRect, moveIndRect);
-
-		int hpIndHeight = 6 * (unit.maxHitPoints <= 5 ? unit.maxHitPoints : 5), hpIndWidth = 6;
-		Rect2 hpIndBackgroundRect = new Rect2(indicatorLoc + new Vector2(-1, 8), new Vector2(hpIndWidth, hpIndHeight));
-		if ((unit.unitType.attack > 0) || (unit.unitType.defense > 0)) {
-			float hpFraction = (float)unit.hitPointsRemaining / unit.maxHitPoints;
-			looseView.DrawRect(hpIndBackgroundRect, Color.Color8(0, 0, 0));
-			float hpHeight = hpFraction * (hpIndHeight - 2);
-			if (hpHeight < 1)
-				hpHeight = 1;
-			var hpContentsRect = new Rect2(hpIndBackgroundRect.Position + new Vector2(1, hpIndHeight - 1 - hpHeight), // position
-										   new Vector2(hpIndWidth - 2, hpHeight)); // size
-			looseView.DrawRect(hpContentsRect, getHPColor(hpFraction));
-			if (unit.isFortified)
-				looseView.DrawRect(hpIndBackgroundRect, white, false);
-		}
-
-		// Draw lines to show that there are more units on this tile
-		if (tile.unitsOnTile.Count > 1) {
-			int lineCount = tile.unitsOnTile.Count;
-			if (lineCount > 5)
-				lineCount = 5;
-			for (int n = 0; n < lineCount; n++) {
-				var lineStart = indicatorLoc + new Vector2(-2, hpIndHeight + 12 + 4 * n);
-				looseView.DrawLine(lineStart                    , lineStart + new Vector2(8, 0), white);
-				looseView.DrawLine(lineStart + new Vector2(0, 1), lineStart + new Vector2(8, 1), Color.Color8(75, 75, 75));
-			}
-		}
-	}
-
-	private static Shader unitShader = null;
-
-	public static Shader getUnitShader()
-	{
-		if (unitShader != null)
-			return unitShader;
-
-		// It would make more sense to use a usampler2D for the indices but that doesn't work. As far as I can tell, (u)int samplers are
-		// broken on Godot because there's no way to create a texture with a compatible format. See:
-		// https://www.khronos.org/opengl/wiki/Sampler_(GLSL)#Sampler_types - Says it's undefined behavior to read from a usampler2d if the
-		// attached texture format is not GL_R8UI.
-		// https://docs.godotengine.org/en/stable/classes/class_image.html#enum-image-format - None of the Godot texture formats correspond to
-		// GL_R8UI. The closest is FORMAT_R8 which corresponds to GL_RED except that won't work since it's a floating point format.
-		string shaderSource = @"
-		shader_type canvas_item;
-
-		uniform sampler2D palette;
-		uniform sampler2D civColorWhitePalette;
-		uniform sampler2D indices;
-		uniform vec2 relSpriteSize; // sprite size relative to the entire sheet
-		uniform vec2 spriteXY; // coordinates of the sprite to be drawn, in number of sprites not pixels
-		uniform vec3 civColor;
-
-		vec4 sampleCivTintedColor(vec2 paletteCoords)
-		{
-			return vec4(civColor, 1.0) * texture(civColorWhitePalette, paletteCoords);
-		}
-
-		void vertex()
-		{
-			UV = (spriteXY + UV) * relSpriteSize;
-		}
-
-		void fragment()
-		{
-			int colorIndex = int(255.0 * texture(indices, UV).r);
-			if (colorIndex >= 254) // indices 254 and 255 are transparent
-				discard;
-			else if (colorIndex >= 240) // indices in [240, 253] are shadows
-				COLOR = vec4(0.0, 0.0, 0.0, float(16 * (255 - colorIndex)) / 255.0);
-			else if (colorIndex >= 224) // indices in [224, 239] are smoke
-				COLOR = vec4(0.0, 0.0, 0.0, 0.1 + 0.032 * float(colorIndex - 224));
-			else {
-				vec2 paletteCoords = vec2(float(colorIndex % 16), float(colorIndex / 16)) / 16.0;
-				bool tintedByCiv = (colorIndex < 16) || ((colorIndex < 64) && (colorIndex % 2 == 0));
-				if (tintedByCiv)
-					COLOR = sampleCivTintedColor(paletteCoords);
-				else
-					COLOR = texture(palette, paletteCoords);
-			}
-		}
-		";
-		var tr = new Shader();
-		tr.Code = shaderSource;
-		unitShader = tr;
-		return tr;
-	}
-
-	private static Shader cursorShader = null;
-
-	public static Shader getCursorShader()
-	{
-		if (cursorShader != null)
-			return cursorShader;
-
-		string shaderSource = @"
-		shader_type canvas_item;
-
-		uniform sampler2D palette;
-		uniform sampler2D indices;
-		uniform vec2 relSpriteSize; // sprite size relative to the entire sheet
-		uniform vec2 spriteXY; // coordinates of the sprite to be drawn, in number of sprites not pixels
-
-		void vertex()
-		{
-			UV = (spriteXY + UV) * relSpriteSize;
-		}
-
-		void fragment()
-		{
-			int colorIndex = int(255.0 * texture(indices, UV).r);
-			if ((colorIndex >= 224) && (colorIndex <= 239))
-				COLOR = vec4(1.0, 1.0, 1.0, float(colorIndex - 224) / float(239 - 224));
-			else if (colorIndex >= 254) // indices 254 and 255 are transparent
-				discard;
-			else {
-				vec2 paletteCoords = vec2(float(colorIndex % 16), float(colorIndex / 16)) / 16.0;
-				COLOR = texture(palette, paletteCoords);
-			}
-		}
-		";
-		var tr = new Shader();
-		tr.Code = shaderSource;
-
-		cursorShader = tr;
-		return tr;
-	}
-}
-
 public partial class BuildingLayer : LooseLayer {
 	private ImageTexture buildingsTex;
 	private Vector2 buildingSpriteSize;
